requirements.txt
django 
pillow
djangorestframework
django_cron
py-moneyed
django-money
django_extensions